feat: state system for handling close handshakes gracefully
We were blindy sending 'normal' closure frames in response
to incoming close frames, regardless of whether of not we initialised
the close. A new state tracker has been added so in the event we
send a close frame, regardless of type, we set the state to 'closing'.

Once we get the response from the peer, we can tell if we started
the shutdown and if so, close the TCP connection, or if the
peer initiated the close, in which case we send the final goodbye.